refactor: Extract c2 calculation logic from PassThroughFilterFactory and cleanup JavaDoc
